Changes on Converter method's names. Fixed error on setting Node's name for debugging geometries.

+++ src/com/jme3/physics/dyn4j/PhysicsSpace.java
@@ -34,8 +34,6 @@ package com.jme3.physics.dyn4j;
 import java.util.List;
 import java.util.logging.Logger;
 
-import javax.vecmath.Vector2d;
-
 import org.dyn4j.collision.Bounds;
 import org.dyn4j.dynamics.Body;
 import org.dyn4j.dynamics.Capacity;
@@ -44,6 +42,7 @@ import org.dyn4j.dynamics.World;
 import org.dyn4j.dynamics.joint.Joint;
 import org.dyn4j.geometry.Vector2;
 
+import com.jme3.math.Vector2f;
 import com.jme3.math.Vector3f;
 
 /**
@@ -98,7 +97,7 @@ public class PhysicsSpace {
         this.physicsWorld = null;
     }
 
-    public void setGravity(final Vector2d gravity) {
+    public void setGravity(final Vector2f gravity) {
         this.physicsWorld.setGravity(new Vector2(gravity.x, gravity.y));
     }
